fix: 공통모달에 picker 적용 1차, 오류수정필요
